Revert "add normalization code"
